+++ changelog.md
@@ -3,6 +3,7 @@
 - Remove unused callback from tests
 - Add loop_safe annotation
 - Remove entity_data_event_callback
+- Make system_event_callback loop aware
 
 # Version 2024.4.5 (2024-04-09)